Moving BulkMutationThrottling functionality into ResourceLimiter. Updating the configuration string for throtting to be more consistent.
